wheels: use manylinux2014 and static libstdc++
- use manylinux2014 with devtoolset-11, which has a recent enough C++
  for our purposes
  - it's not on ARM but we're not building for ARM Linux anyway
- statically link libstdc++ on linux to avoid runtime ld errors

+++ wheel_build/before_all_linux.sh
@@ -6,9 +6,10 @@ if command -v apk; then
         tcl-dev readline-dev tcl-readline-dev
 fi
 if command -v yum; then
-    yum install -y swig flex zlib-devel readline-devel eigen3-devel \
+    yum remove -y swig || true # can be installed by default
+    yum install -y swig3 flex zlib-devel eigen3-devel \
         elfutils-devel elfutils-libelf-devel libdwarf-devel binutils-devel m4 \
-        perl-core tcl-devel tcl-tclreadline-devel
+        perl-core tcl-devel devtoolset-11
 fi
 
 NPROC=$(getconf _NPROCESSORS_ONLN 2>/dev/null || sysctl -n hw.ncpu)
@@ -32,5 +33,24 @@ if ! printf '%s\n' '%require "3.8"' '%%' 'start: ;' | bison -o /dev/null /dev/st
         make clean
         $SUDO make install -j$NPROC
     )
-rm -rf "$WORKDIR"
+    rm -rf "$WORKDIR"
+fi
+
+# tcl out of date in CentOS 7 packages
+TCL_VERSION=8.6.16
+TCL_HASH=05c061cca79a17efc61ef7c1cd873227a02a7c1fdc114aa6a4ecdd33ecf167f4
+if ! echo 'exit [expr [package vcompare [info patchlevel] 8.6] < 0]' | tclsh ; then
+    WORKDIR=$(mktemp -d)
+    (
+        cd $WORKDIR
+        curl -L --retry 5 --retry-delay 3 \
+            "https://github.com/tcltk/tcl/archive/refs/tags/core-$(echo $TCL_VERSION | tr '.' '-').tar.gz" > tcl.tgz
+        echo "$TCL_HASH tcl.tgz" | sha256sum -c
+        tar --strip-components=1 -xzC . -f tcl.tgz
+        ./unix/configure
+        make -j$NPROC
+        $SUDO make install -j$NPROC
+        $SUDO ln -s /usr/local/bin/tclsh8.6 /usr/local/bin/tclsh
+    )
+    rm -rf "$WORKDIR"
 fi
